crypto/requirements.txt
# Core
pydantic>=2.0
pydantic-settings>=2.0
pydantic-ai>=0.1
structlog>=24.0
httpx>=0.27

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13

# Redis
redis>=5.0

# Alpaca
alpaca-py>=0.28

# Technical Analysis
numpy>=1.26
pandas>=2.2

# Telegram
python-telegram-bot>=21.0

# Scheduling
apscheduler>=3.10

# AI
anthropic>=0.40

# Terminal UI
rich>=13.0

# Web Dashboard
fastapi>=0.115
uvicorn[standard]>=0.32
python-multipart>=0.0.9

# Testing
pytest>=8.0
pytest-asyncio>=0.24
